♻️ refactor: clarify search document Zod schema naming
