+++ packages/database/src/repositories/ftsSearchDocument/index.ts
@@ -1,4 +1,4 @@
 export * from './builder';
 export * from './fingerprint';
 export * from './mappings';
-export * from './schema';
+export * from './zodSchema';
