+++ packages/database/src/repositories/ftsSearchDocument/zodSchema.ts
